+++ src/App.js
@@ -1,6 +1,6 @@
 //components
 import Main from "./components/Main";
-import Header from "./components/Header";
+import Header from "./components/header/Header";
 import { useState } from "react";
 
 function App() {
