+++ src/styles/header.css
@@ -1,12 +1,11 @@
 #header{
-    border-bottom: 3px rgb(37, 34, 34) solid;
     position: absolute;
     width: 100vw;
     top: 0;
     z-index: 999;
     background-color: white;
     overflow: hidden;
-    box-shadow: 0px 0px 15px rgb(226, 223, 223);
+    box-shadow: 0px 0px 10px rgb(170, 168, 168);
 }
 
 .closedMenu{
@@ -95,19 +94,18 @@
 
 .cartItem{
     display: flex;
-    min-width: 18%;
+    flex-direction: column;
+    min-width: 5em;
+    max-width: 5em;
     /* height: 80%; */
     /* background-color: aqua */
 }
 .cartItem img{
-    height: 100%;
+    width: 100%;
 }
 
 .cartItemDesc{
-    /* margin: 0; */
-}
-.cartItemDesc h5{
-    margin: 2px;
+    display: flex;
 }
 
 @media only screen and (max-width: 600px){
